safari-600.7-branch/LayoutTests/svg/animations/animate-path-to-animation-expected.txt

SVG 1.1 dynamic animation tests

Test calcMode spline with to animation. You should see a green 100x100 path and only PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS path.animatedPathSegList.numberOfItems is 5
PASS path.animatedPathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.animatedPathSegList.getItem(0).x is 40
PASS path.animatedPathSegList.getItem(0).y is 40
PASS path.animatedPathSegList.getItem(1).pathSegTypeAsLetter is "L"
PASS path.animatedPathSegList.getItem(1).x is 60
PASS path.animatedPathSegList.getItem(1).y is 40
PASS path.animatedPathSegList.getItem(2).pathSegTypeAsLetter is "L"
PASS path.animatedPathSegList.getItem(2).x is 60
PASS path.animatedPathSegList.getItem(2).y is 60
PASS path.animatedPathSegList.getItem(3).pathSegTypeAsLetter is "L"
PASS path.animatedPathSegList.getItem(3).x is 40
PASS path.animatedPathSegList.getItem(3).y is 60
PASS path.animatedPathSegList.getItem(4).pathSegTypeAsLetter is "Z"
PASS path.pathSegList.numberOfItems is 5
PASS path.pathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.pathSegList.getItem(0).x is 40
PASS path.pathSegList.getItem(0).y is 40
PASS path.pathSegList.getItem(1).pathSegTypeAsLetter is "L"
PASS path.pathSegList.getItem(1).x is 60
PASS path.pathSegList.getItem(1).y is 40
PASS path.pathSegList.getItem(2).pathSegTypeAsLetter is "L"
PASS path.pathSegList.getItem(2).x is 60
PASS path.pathSegList.getItem(2).y is 60
PASS path.pathSegList.getItem(3).pathSegTypeAsLetter is "L"
PASS path.pathSegList.getItem(3).x is 40
PASS path.pathSegList.getItem(3).y is 60
PASS path.pathSegList.getItem(4).pathSegTypeAsLetter is "Z"
PASS path.animatedPathSegList.numberOfItems is 5
PASS path.animatedPathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.animatedPathSegList.getItem(0).x is 20
PASS path.animatedPathSegList.getItem(0).y is 20
PASS path.animatedPathSegList.getItem(1).pathSegTypeAsLetter is "L"
PASS path.animatedPathSegList.getItem(1).x is 80
PASS path.animatedPathSegList.getItem(1).y is 20
PASS path.animatedPathSegList.getItem(2).pathSegTypeAsLetter is "L"
PASS path.animatedPathSegList.getItem(2).x is 80
PASS path.animatedPathSegList.getItem(2).y is 80
PASS path.animatedPathSegList.getItem(3).pathSegTypeAsLetter is "L"
PASS path.animatedPathSegList.getItem(3).x is 20
PASS path.animatedPathSegList.getItem(3).y is 80
PASS path.animatedPathSegList.getItem(4).pathSegTypeAsLetter is "Z"
PASS path.pathSegList.numberOfItems is 5
PASS path.pathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.pathSegList.getItem(0).x is 40
PASS path.pathSegList.getItem(0).y is 40
PASS path.pathSegList.getItem(1).pathSegTypeAsLetter is "L"
PASS path.pathSegList.getItem(1).x is 60
PASS path.pathSegList.getItem(1).y is 40
PASS path.pathSegList.getItem(2).pathSegTypeAsLetter is "L"
PASS path.pathSegList.getItem(2).x is 60
PASS path.pathSegList.getItem(2).y is 60
PASS path.pathSegList.getItem(3).pathSegTypeAsLetter is "L"
PASS path.pathSegList.getItem(3).x is 40
PASS path.pathSegList.getItem(3).y is 60
PASS path.pathSegList.getItem(4).pathSegTypeAsLetter is "Z"
PASS path.animatedPathSegList.numberOfItems is 5
PASS path.animatedPathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.animatedPathSegList.getItem(0).x is 0
PASS path.animatedPathSegList.getItem(0).y is 0
PASS path.animatedPathSegList.getItem(1).pathSegTypeAsLetter is "L"
PASS path.animatedPathSegList.getItem(1).x is 100
PASS path.animatedPathSegList.getItem(1).y is 0
PASS path.animatedPathSegList.getItem(2).pathSegTypeAsLetter is "L"
PASS path.animatedPathSegList.getItem(2).x is 100
PASS path.animatedPathSegList.getItem(2).y is 100
PASS path.animatedPathSegList.getItem(3).pathSegTypeAsLetter is "L"
PASS path.animatedPathSegList.getItem(3).x is 0
PASS path.animatedPathSegList.getItem(3).y is 100
PASS path.animatedPathSegList.getItem(4).pathSegTypeAsLetter is "Z"
PASS path.pathSegList.numberOfItems is 5
PASS path.pathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.pathSegList.getItem(0).x is 40
PASS path.pathSegList.getItem(0).y is 40
PASS path.pathSegList.getItem(1).pathSegTypeAsLetter is "L"
PASS path.pathSegList.getItem(1).x is 60
PASS path.pathSegList.getItem(1).y is 40
PASS path.pathSegList.getItem(2).pathSegTypeAsLetter is "L"
PASS path.pathSegList.getItem(2).x is 60
PASS path.pathSegList.getItem(2).y is 60
PASS path.pathSegList.getItem(3).pathSegTypeAsLetter is "L"
PASS path.pathSegList.getItem(3).x is 40
PASS path.pathSegList.getItem(3).y is 60
PASS path.pathSegList.getItem(4).pathSegTypeAsLetter is "Z"
PASS path.animatedPathSegList.numberOfItems is 5
PASS path.animatedPathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.animatedPathSegList.getItem(0).x is 40
PASS path.animatedPathSegList.getItem(0).y is 40
PASS path.animatedPathSegList.getItem(1).pathSegTypeAsLetter is "L"
PASS path.animatedPathSegList.getItem(1).x is 60
PASS path.animatedPathSegList.getItem(1).y is 40
PASS path.animatedPathSegList.getItem(2).pathSegTypeAsLetter is "L"
PASS path.animatedPathSegList.getItem(2).x is 60
PASS path.animatedPathSegList.getItem(2).y is 60
PASS path.animatedPathSegList.getItem(3).pathSegTypeAsLetter is "L"
PASS path.animatedPathSegList.getItem(3).x is 40
PASS path.animatedPathSegList.getItem(3).y is 60
PASS path.animatedPathSegList.getItem(4).pathSegTypeAsLetter is "Z"
PASS path.pathSegList.numberOfItems is 5
PASS path.pathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.pathSegList.getItem(0).x is 40
PASS path.pathSegList.getItem(0).y is 40
PASS path.pathSegList.getItem(1).pathSegTypeAsLetter is "L"
PASS path.pathSegList.getItem(1).x is 60
PASS path.pathSegList.getItem(1).y is 40
PASS path.pathSegList.getItem(2).pathSegTypeAsLetter is "L"
PASS path.pathSegList.getItem(2).x is 60
PASS path.pathSegList.getItem(2).y is 60
PASS path.pathSegList.getItem(3).pathSegTypeAsLetter is "L"
PASS path.pathSegList.getItem(3).x is 40
PASS path.pathSegList.getItem(3).y is 60
PASS path.pathSegList.getItem(4).pathSegTypeAsLetter is "Z"
PASS successfullyParsed is true

TEST COMPLETE
